move nearestParentWorkspace into workspace_util
